GH-29: :building_construction: :truck: Rename `PatientRecordsSearch` to `PatientSearchForm` and refactor implementation
