+++ app/screens/patientRecords/PatientSearchForm.stories.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import PatientSearchForm from './PatientSearchForm';
+
+storiesOf('Screens/PatientRecords/PatientSearchForm', module).add(
+  'default',
+  () => <PatientSearchForm />
+);
